refactor login codes
